LayoutTests/accessibility/focusable-inside-hidden-expected.txt

This tests that a focusable object inside an aria-hidden is in the hieararchy only after it gains focus.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS !button || !button.isValid is true
PASS button.description is 'AXDescription: BUTTON'
PASS button.description is 'AXDescription: BUTTON'
PASS !button || !button.isValid is true
PASS successfullyParsed is true

TEST COMPLETE
a
test
